+++ tests/functional/test_linspace_grad.py
@@ -6,25 +6,25 @@ from torchoptics.functional import linspace_grad
 def test_output_values():
     start_val, end_val = 0.0, 1.0
     steps = 10
-    start = torch.tensor(start_val, requires_grad=True, dtype=torch.float64)
-    end = torch.tensor(end_val, requires_grad=True, dtype=torch.float64)
-    expected = torch.linspace(start_val, end_val, steps, dtype=torch.float64)
+    start = torch.tensor(start_val, requires_grad=True)
+    end = torch.tensor(end_val, requires_grad=True)
+    expected = torch.linspace(start_val, end_val, steps)
     actual = linspace_grad(start, end, steps)
     assert torch.allclose(actual, expected)
 
 
 def test_edge_cases():
     for start_val, end_val, steps in [(1, 1, 0), (1, 3, 0), (12, 12, 1), (12, 14, 1)]:
-        start = torch.tensor(start_val, requires_grad=True, dtype=torch.float64)
-        end = torch.tensor(end_val, requires_grad=True, dtype=torch.float64)
-        expected = torch.linspace(start_val, end_val, steps, dtype=torch.float64)
+        start = torch.tensor(start_val, requires_grad=True)
+        end = torch.tensor(end_val, requires_grad=True)
+        expected = torch.linspace(start_val, end_val, steps)
         actual = linspace_grad(start, end, steps)
         assert torch.allclose(actual, expected)
 
 
 def test_differentiability():
-    start = torch.tensor(-3, requires_grad=True, dtype=torch.float64)
-    end = torch.tensor(1.0, requires_grad=True, dtype=torch.float64)
+    start = torch.tensor(-3, requires_grad=True)
+    end = torch.tensor(1.0, requires_grad=True)
     steps = 10
     linspace_tensor = linspace_grad(start, end, steps)
     output = linspace_tensor.sum()
